pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marvin"
description = "Multi-model Agentic Reviewers for Virtually Intelligent Notations"
readme = "README.md"
version = "0.0.1"
license = "APACHE-2.0"
authors = [
    {name = "Jeffrey Clark", email = "h0tw1r3@gmail.com"},
    {name = "Nikita Filonov", email = "nikita.filonov@gmail.com"}
]
keywords = [
    "ai",
    "code review",
    "llm",
    "openai",
    "claude",
    "gemini",
    "ollama",
    "bedrock",
    "openrouter",
    "ci/cd",
    "gitlab",
    "github",
    "gitea",
    "bitbucket",
    "bitbucket cloud",
    "bitbucket server",
    "azure",
    "azure devops",
    "azure openai"
]
maintainers = [
    {name = "Jeffrey Clark", email = "h0tw1r3@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]
dependencies = [
    "typer",
    "httpx",
    "pyyaml",
    "loguru",
    "aiofiles",
    "pydantic",
    "pydantic-settings"
]
requires-python = ">=3.12"

[project.urls]
Issues = "https://github.com/h0tw1r3/marvin/issues"
Homepage = "https://github.com/h0tw1r3/marvin"
Repository = "https://github.com/h0tw1r3/marvin"

[project.scripts]
marvin = "marvin.cli.main:app"

[dependency-groups]
lint = [
    "ruff"
]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-env",
    "coverage"
]
dev = [
    {include-group = "lint"},
    {include-group = "test"}
]

[tool.uv]
package = true
python-preference = "system"

[tool.coverage.report]
ignore_errors = true
show_missing = true
fail_under = 89

[tool.pytest.ini_options]
testpaths = [
    "tests/**/test_*.py"
]
addopts = [
    "--no-header",
    "--tb=short",
    "--cov=src/marvin",
    "--cov-branch",
    "--cov-report=term-missing"
]

[tool.pytest_env]
MARVIN_CONFIG_FILE_YAML = "tests/configs/config-test.yaml"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "TC"]
ignore = [
    "I001"  # isort found an import in the wrong position (handled by isort)
]

[tool.pymarkdown]
extensions.markdown-tables.enabled = true

[tool.pymarkdown.plugins.md013]
enabled = true
code_blocks = false
line_length = 160

[tool.pymarkdown.plugins.per-file-ignores]
"docs/troubleshooting/*.md" = "MD024"
"docs/prompts/**/*.md" = "MD001,MD004,MD007,MD013,MD026"
"src/marvin/prompts/*.md" = "MD001,MD007,MD041"
